Порефакторил компонент trip-controller, вынес EmptyEvent и ViewMode в constants.js
